pyproject.toml
[tool.poetry]
name = "traffic-signs-detection"
version = "0.1.0"
description = "Simple traffic signs detector based on YOLOv8n"
authors = ["Ilya Tsarin <tsarin.ii@phystech.edu>"]

[tool.poetry.dependencies]
python = "~3.10"
ultralytics = "^8.1.36"
fire = "^0.6.0"
lightning = "^2.2.1"
hydra-core = "^1.3.2"
easydict = "^1.13"
mlflow = "^2.12.1"
dvc-gdrive = "^3.0.1"
pillow = "^10.3.0"
torchvision = "^0.18.0"
torch = "^2.0"
onnx = "^1.16.2"
onnxscript = "^0.1.0.dev20240925"
tensorrt = "^10.4.0"
pycuda = "^2024.1.2"

[tool.poetry.group.dev.dependencies]
pre-commit = "^3.6.0"

[tool.black]
line-length = 90
target-version = ["py310"]

[tool.isort]
src_paths = ["detector"]
profile = "black"
line_length = 90
lines_after_imports = 2

[tool.flake8]
max-line-length = 90
ignore = ['E203', 'E501', 'W503', 'B950']
max-complexity = 12
select = ['B', 'C', 'E', 'F', 'W', 'B9']
per-file-ignores = [
    # for easier imports to __init__ without __all__
    '**/__init__.py: F401',
]
count = true

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
